discord_receipt_bot/requirements.txt
discord.py
jinja2
pdfkit
pyemail